extract hierarchical_sub_task_graph_builder_util
